+++ ValidationProject/Assets/Editor/ApexValidationProjectEntry.cs
@@ -1,0 +1,18 @@
+#if UNITY_EDITOR
+using DAZI.Apex.Tools;
+using UnityEditor;
+
+public static class ApexValidationProjectEntry
+{
+    [MenuItem("Apex Validation/Build Scene And Validate")]
+    public static void BuildSceneAndValidate()
+    {
+        ApexValidationSceneBuilder.BuildAndValidateBatch();
+    }
+
+    public static void RunBatch()
+    {
+        ApexValidationSceneBuilder.BuildAndValidateBatch();
+    }
+}
+#endif